backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
openai==1.3.7
yfinance==0.2.28
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
pydantic-settings==2.1.0
supabase==2.18.1